VV dropdown proccall update.
This is the bit of code that is called by going for `Advanced ProcCall` in the VV dropdown menu. Previously, `callproc(D)` was the procedure for calling something directly on said D. With this PR's renaming, `callproc()` is the general Advanced ProcCall for anything - this needs to be replaced with `callproc_datum(D)` to retain the functionality of going straight for procs of the datum being VVed.