Research Nerfs (#1863)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

This PR brings in some changes to cut down a few things.
one monkey cube box has been removed to cut down the maximum xenomorphs
researchers can make (from 21 xenomorphs total to 11, excluding CL
monkey cubes and including pooled).

A corrupted hive's biggest advantage over the regular hive is that they
are teamed up with the marines, so if a corrupted hive managed to get
ALL the monkey cubes for their hive, they would effectively be 1:1 with
the rival hive with the addition of being allied to the marines, an USCM
allied hive still has the advantage over the regular hive by being
allied to marines and being able to plan out their castes for XvX than
XvH.

This change will now make the regular hive atleast have a chance
fighting off the corrupted hive in a scenario with marines and be
slightly better when the corrupted are alone (if the regular hive has
more xenos).

If you are upset that you lost more monkeys to "experiment" on that isnt
just infecting them with huggers, keep in mind that there is no way you
can permamently lose a monkey unless you decapitate them or leave them
for dead instead of reviving them. No I will not make special monkeys
that cannot be infected, use your monkeys as intended for experiments,
they are not only for making a corrupted hive.

second change is increasing the rarity of Musclestimulating (MST) and
Nervestimulating (NST) from Uncommon to Rare (1 tier up). Lately if not
for a very long time, Research seems to only ever make stimulants, we
have cut down a lot of ways to get them (by removing Synaptizine and
Hyperzine and other changes etc.) and currently you can still obtain
them quite easily and the whole research meta was to always make NST MST
instead of utilizing literally any other beneficial property and now
with anti xenomorph interactions, they still make stimulants only
because it has been ingrained in their brain for too long.

Increasing their rarity makes the properties no longer appear in botany
chems, so you will now have to wait for colony vials/papers, a xeno
corpse and max your clearance or buy a C3+ paper.

The final change is replacing the Ultrazine Pill Bottle the CL has with
an Ultrazine Auto-Injector, the pill bottle contained 5 pills with 5u
each, the new injector contains 25u of the stimulant and gives 5u each
injection. The only difference being is that the Ultrazine the CL has is
now in a liquid injection than pill form, it will begin affecting you
instantly and the main reason I made this change, you can no longer
obtain the chemical from it as you could with pills by dissolving them
which granted an easy scan for MST, since we are making MST rarer, the
CL should not have pills but an auto-injector.


# Explain why it's good for the game

Cuts down corrupted numbers to prevent easy stomps whenever marines get
them
An attempt to change the stim meta (by making it rarer) to encourage
other properties to be used.


# Testing Photographs and Procedure
Checked on using the briefcase to spawn the new autoinjector to make
sure it spawns, and tested it to make sure it has the correct values.
Checked that the removed monkey cube box is actually gone. Checked that
MST and NST are now actually Rare by scanning them and seeing their
rarity type in the simulator.


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl: Unknownity
del: Removed one monkey cube box from Research.
balance: Increased rarity of MST and NST from Uncommon to Rare.
balance: Replaced the Ultrazine Pill Bottle with an Ultrazine
auto-injector in the CL Vendor Briefcase.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

Co-authored-by: Unknownity <a>
Co-authored-by: Stan_Albatross <66756236+stanalbatross@users.noreply.github.com>