Smartfridge tgui (#1623)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Migrates the smart fridge to TGUI and reworks internal logic.

Adds a standard component to use for displaying an electrical access
panel for hacking. The goal is to indicate that there is an access panel
and to try and make the UI more immersive.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Smartfridge with some items loaded

![image](https://user-images.githubusercontent.com/101871009/202933478-2c1df4d2-970c-4fbb-b9aa-ba590046aadf.png)

Smartfridge with panel open and wire status lights active. Solid green
for good or flashing red for bad.

![image](https://user-images.githubusercontent.com/101871009/202933496-03b22590-b8a8-4ef4-89d7-30c63a941fb3.png)


<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

## Changelog

<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
fix: fixed bug in turing autodispensor so that pill bottles can be in
local smart fridge
refactor: migrated smart fridge to tgui
refactor: added common electrical panel for tgui components
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->