Flares visual rework (#1877)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request


![image](https://user-images.githubusercontent.com/604624/206264570-f4022249-80f2-4ef9-9f53-4eb68dd44cf7.png)

This:
 * Makes flares land randomly rotated after they spin in throwing
* Reworks sprites to separate the flame burn animation and so it's
always facing north side
* Converts flame burn to a greyscale animation using multiply+additive
to achieve similar looks in any color

So at the end of the day, flares of arbitrary colors land randomly with
their flame facing up.

Some explanations and caveats:
* For the "flame always up" part, the simpler way of doing this was to
fit the flares into half the spritesheet, to have their lighting point
in center of it. This means it rotates like the hands of a clock, and
the flame can always be put in the middle. Otherwise, this requires
snowflake in-code calculations to position it right.
* As a result, alignment is approximative (one pixel off) depending on
rotation.
* To achieve this, 2 lighting pixels had to be cropped from Signal
Flare's back.
* To achieve this, the flame animation had to be reduced from about
19-21 pixels long to roughly 18 to fit into frame. I'm not entirely
happy about this.
* You can't straight up greyscale the flame, because it's not "equally"
as "red" as it is "white" (consider RGB for the 3 main pixels of old
flare burn: (239,71,63), (255,119,101), (255,172,161)). To get around
this, we apply it twice: once as traditional greyscale tint for the
flame gradient, and a second time as additive mask with the broader
flame color. This is probably wrong as I have no idea what i'm doing,
but it seems to work overall....
* It might make sense to have a bigger gap in the gray scale as a
result. I've never done this, i dunno!

# Explain why it's good for the game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

It's just cool visually and helps anyone wanting to use colored flares
in the future.

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
See above

# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
add: Flares will now land in different positions as they spin from
throwing.
add: Flares flame is now separate and will always face up.
add: Converted flares flames to grayscale, allowing flares of any color
to be made.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->