Reduces the volume of M44 and MOU53's firing sounds. (#1948)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

The M44 revolver and MOU53 break action shotgun are too loud when fired,
so I lowered their sounds to be roughly the same as the M37 shotgun,
which is a loud enough weapon, considering the fact that it is a
shotgun.

# Explain why it's good for the game

I like the M44 revolver and MOU53 break action shotgun, but I wear
headphones and they're so loud that they're contributing towards my
hearing loss.

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>




![image](https://user-images.githubusercontent.com/16822632/207688309-80f1b877-521f-43df-9223-17694ec751cd.png)
A screenshot of the M37's volume in Audacity. It's just over 12 dB. I
used this as a reference.


![image](https://user-images.githubusercontent.com/16822632/207689061-1094033b-9e94-42b7-be68-6532e650d901.png)
A screenshot of teh MOU53's volume. It's much louder.


![image](https://user-images.githubusercontent.com/16822632/207689370-0c083f78-f3e3-48e4-98ab-3de18aa4968c.png)
Same with the M44.



![image](https://user-images.githubusercontent.com/16822632/207689992-9650df7e-76dc-4687-8fcd-806c6dab7bf8.png)
New M44.


![image](https://user-images.githubusercontent.com/16822632/207690310-97576977-7ed9-4fb1-8078-5116e55d149b.png)
New MOU.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl: Lacasitospro
fix: Made the M44 and MOU53 less loud.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->