Added a super snowflake vendor for devs and admins to play with (#1648)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

I was doing some work on picking a set of clothes for a fun ERT when I
realised we could use a vendor for admins / devs to figure out what kind
of items look good together and so on before they spawn a bunch of mobs
with them. So I added a few Super Snowflake Vendors that have all the
items of a given type in them. The UI is a bit slow due to the number of
items, and when I tried putting all of the items together into one
vendor it lagged out the game, so I split them up into separate types.

With some feedback from carlarc, I also made the Super Snowflake vendor
be easily customizable - if you spawn the base variant it will ask you
to provide an item path to stock. You can also easily configure it while
mapping to stock a given set of subitems.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

It makes it easier to match interesting outfit combinations for ERTs and
what have you. This can let admins communicate more clearly what they
might want in a gear preset, and for devs to easier test how things look
together.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

## Changelog

<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
add: Added a set of Super Snowflake Vendors for admins and devs to play
with. If you spawn the base one you can stock it with any item and its
subtypes as defined when you first spawn it / interact with it.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->