Fix Prisoner will now be able to witraw money at brig. (#1938)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request
add an ATM into brig so that MP can retrieve cash for prisoner to use in
jail.
moved a medical vendor east of where ATM is.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Able prisoner to gain acces to cash.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>

Before

![image](https://user-images.githubusercontent.com/117036822/207556284-91b9e9ba-c5d3-4ea0-bb85-817ecdd3317e.png)
After

![image](https://user-images.githubusercontent.com/117036822/207554743-288a67b2-7b5a-448b-8842-f6995d1ecbe0.png)

<summary>Screenshots & Videos</summary>



Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
add: Added an ATM in brig and move the med vendor in brig infirmary 
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

Co-authored-by: Julien <jverger.ingx@gmail.com>