[MDB IGNORE] Removal of groundside nanomed refilling splints and skilless injectors. Nerfs non-med APC nanomeds. (#1758)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Removes nanomed refilling splints and skilless injectors.

## Why It's Good For The Game

This is step one of an overall medical nerf for the average PFC. The
ability to refill the injectors on any nanomed in the colony is a bit
too powerful all on top of making med APC less important as a source of
medical refill.

## Changelog

:cl: Morrow
balance: Removes groundside nanomeds refilling splints and skilless
injectors.
balance: Nerfs non-med APC nanomeds to basic groundside level.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->