Makes SMESes actually fixable (#1977)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

Currently the broken SMESes used in certain maps don't charge even when
that have been fully repaired, this aims to fix that.
Also, for some reason the terminals appeared on the opposite side than
expected so this fixes that aswell.

# Explain why it's good for the game

Broken SMESes being unfixable made it so on certain maps (read: Fiorina)
the power objective was impossible, objectives being completable and
things working as expected is good.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

I repaired a broken SMES and checked if it charged, I also checked if
the power objective counted it as outputting, which it did.
Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl: CometBlaze
fix: Terminals will now spawn on the correct side of SMESes
fix: Round start broken SMESes will now function as expected once
repaired
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->