MP Cadet Removal (#1993)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

This is part of an atomization of #1319.

This PR removes MP Cadet. Instead playing as an MP will make you a LCPL
for the first 10 hours and then you will be a full sergeant as is the
current case with nurses.

# Explain why it's good for the game

MP Cadets were never meant to be just more MP slots.

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>
Yup, I tested it.

</details>


# Changelog

:cl: Morrow
add: MPs will now start at LCPL for the first 10 hours of playtime.
del: MP Cadet is removed.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->