Fixes RP by adding the operation name to marines status tab (#2021)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request
This adds the operation name, which can only be seen at first launch of
the dropship and at the end of the round, to the status tab. It's only
available to the one affiliated to the marine faction since xenos
shouldn't know that and ghosts are LRP (Decided by yours truly)

**Tested on WO, didn't cause any runtimes that weren't already there** 👍
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Spartanbobby (Head of RP) mentioned that idea once and no one did it, so
I'm sacrificing my GBP for it. I think it adds alot of discussion
possibilities, the XO/CO can mention it during the briefing and it's not
something that pollute the status tab too much since it's already not
used for much. See pictures below.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>


![opname](https://user-images.githubusercontent.com/53326609/208552587-36396fe9-fb21-4123-acec-6e4e3524cee9.png)


![opnamemore](https://user-images.githubusercontent.com/53326609/208552595-9ec9697e-b053-440f-9544-9153693f2e6c.png)


Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl: Molto
add: Added the operation name to the marines status tab
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->