rename lethal mode on smart weaponry to IFF (#2049)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

renames lethal mode on the M46C and the M56 to IFF

# Explain why it's good for the game

 its less confusing for newer players


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Stalkerino
fix: renamed lethal mode to IFF on the M46C and the M56
/:cl: