Re-added the Rescue Survivors Objective (#1613)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

I noticed the Rescue Survivors Objective was not around anymore. After
checking on Discord whether it was removed on purpose (apparently it was
not), I re-added it:


![image](https://user-images.githubusercontent.com/964559/202845431-34448002-7ef7-4f68-bab3-044c47bbab18.png)

(default survivors net you 0.7 points)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

The objective was a neat way to incentivise Survivors to survive for as
long as possible and for the Almayer crew not to dilly dally, deploy
early and sweep the colony for survivors or their revivable corpses.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

## Changelog

<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
add: Re-added the old objective of rescuing survivors. Now each survivor
rescued alive and brought to Almayer will net the USCM a good deal of
points for their Techweb.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->