Tweak: Boiler bombard cooldown (#2187)
# About the pull request

The cooldown for the bombard ability is applied before the shot actually
fires. If the boiler moves / uses plasma / has no line-of-sight after
the 1.5 second wind-up, the shot will not fire, but the cooldown will
still apply. This makes no sense personally.

This PR simply makes it so the cooldown only applies if the shot
actually fires. Note the cooldown duration is adjusted to compensate for
the cooldown now being applied **after** the 1.5 second wind-up.

# Explain why it's good for the game

Getting put on cooldown for an ability that never actually got used
seems like a bug?

# Testing Photographs and Procedure

Even with the delayed cooldown application, the ability still behaves
like normal.

# Changelog

:cl: Casper
fix: boiler bombard cooldown now only applies if the shot actually fires
/:cl: