Mapping Add back the ability to open secure dome. (#1887)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request
Long time ago survivors could open secure storage with a button in
corporate dome.
The idea is to reimplement the possibility of opening but in a more
complexe way.
First now Secure storage as is two blast door seperate Outer and inner.
location for the button for outer blast door are corporate office,
Commander office.
Location for inner blast door button are in research and Armory in
nexus.
Survivors will need to coordinate for this to work.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Make LV-624 more coherent by adding a possibility to get in secure
storage without breaking the walls...

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>


<summary>Screenshots & Videos</summary>


![image](https://user-images.githubusercontent.com/117036822/206397346-e7737e3f-d4a5-42a4-b87b-3f1e002c5e46.png)


![image](https://user-images.githubusercontent.com/117036822/206397530-9be7d9e0-b8a4-433a-8059-d2256bb0da46.png)


![image](https://user-images.githubusercontent.com/117036822/206397635-06662d2d-9dbb-468e-9fcb-4d0cde595c0c.png)


![image](https://user-images.githubusercontent.com/117036822/206397738-935400e4-8c64-4055-9306-5f33cd41abb8.png)


![image](https://user-images.githubusercontent.com/117036822/206397839-b21e999c-9faa-414d-81d2-fdf1c1409bd6.png)


Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
add: add buttons for Inner and Outer Blast door of secure storage for
you to find out where.
del: deleted Nexus lockdown and garage shutter button in armory but it's
still in mapping data.
balance: adding the ability for survivor to get in secure storage on
LV-624.
code: separated Inner and outer blast door instead of being a single
entity.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

Co-authored-by: Julien <jverger.ingx@gmail.com>