Colony Revolver Adjustments (#2116)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Changed the Smith & Wesson's caliber from .357 to .38 to remove
compatibility between its rounds and the Spearhead revolver's.

Reduced the amount of time the Smith & Wesson's trick buff lasts.

The CMB Spearhead revolver can no longer fire in bursts nor fit the
burst-fire assembly.

# Explain why it's good for the game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

> Changed the Smith & Wesson's caliber from .357 to .38 to remove
compatibility between its rounds and the Spearhead revolver's.

This can be exploited to put hollowpoint rounds in the S&W, which
combined with its damage multiplier will deal a truly ridiculous amount
of damage, 2-shooting runners in close range.

I strongly believe the multiplier is not the problem - it's the damage.
And the hollowpoint ammo's damage is doubled by the multiplier, while
its intended 'weaker against armor' downside isn't multiplied.

> Reduced the amount of time the Smith & Wesson's trick buff lasts.

It was 8 seconds, i have no idea how I put that value in, maybe it was
for testing purposes and i forgot to remove it. Having it be so long
kind of negates the gimmick of doing a trick and then firing the gun
since it stays active for so long.

> The CMB Spearhead revolver can no longer fire in bursts nor fit the
burst-fire assembly.

I kept this in because I thought it fit the lore - turns out it doesn't!
It's stupid and makes no sense in any way so goodbye legacy poopcode.

These buffs should keep these revolvers' strong utility but ensure no
cheesy overcomplicated strats can be used to instantly satanize Xenos
with. (Or your own teammate..)

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Be sure to properly mark your PRs to prevent unnecessary
GBP loss. Maintainers freely reserve the right to remove and add tags
should they deem it appropriate. -->

:cl:
balance: Changed the Smith & Wesson's caliber from .357 to .38 to remove
compatibility between its rounds and the Spearhead revolver's.
balance: Reduced the amount of time the Smith & Wesson's trick buff
lasts.
del: The CMB Spearhead revolver can no longer fire in bursts nor fit the
burst-fire assembly.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

Co-authored-by: Stan_Albatross <66756236+stanalbatross@users.noreply.github.com>